NIFIREG-292 Introduce IdentityMapper interface and remove direct use of IdentityMappingUtil and NiFiRegistryProperties from user-group and policy providers, collapse abstract database providers into implementations
