+++ nifi-registry-core/nifi-registry-framework/src/main/java/org/apache/nifi/registry/security/identity/IdentityMapper.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.nifi.registry.security.identity;
+
+
+public interface IdentityMapper {
+
+    /**
+     * Maps the given user identity to a target identity.
+     * If the given identity does not match configured patters, the input is returned.
+     *
+     * @param userIdentity the identity to map
+     * @return the mapped identity, or the same identity if no mappings matched
+     */
+    String mapUser(String userIdentity);
+
+    /**
+     * Maps the given group name to a target group name.
+     * If the given group name does not match configured patters, the input is returned.
+     *
+     * @param groupName the group name to map
+     * @return the mapped group name, or the same group name if no mappings matched
+     */
+    String mapGroup(String groupName);
+
+}
